Edit: views and services in account, meta, exchange domains. Migrate done

+++ accounts/application/dto.py
@@ -1,9 +1,9 @@
 from dataclasses import dataclass
-from typing import List
+from decimal import Decimal
 
 @dataclass(frozen=True)
 class AccountUpdateCommand:
     country_code: str
     currency: str
-    annual_income: int
-    monthly_investable_amount: int
+    annual_income: Decimal
+    monthly_investable_amount: Decimal
